feat(spoiler): define media aware store
