+++ projects/client/src/lib/features/spoilers/_internal/useSpoiler.ts
@@ -5,13 +5,15 @@ import { derived } from 'svelte/store';
 export function useSpoiler() {
   const { isAuthorized } = useAuth();
 
-  return {
-    isSpoilerHidden: derived(isAuthorized, ($isAuthorized) => {
-      if (!$isAuthorized) {
-        return false;
-      }
+  const isSpoilerHidden = derived(isAuthorized, ($isAuthorized) => {
+    if (!$isAuthorized) {
+      return false;
+    }
 
-      return Boolean(useUser().current().preferences.isSpoilerHidden);
-    }),
+    return Boolean(useUser().current().preferences.isSpoilerHidden);
+  });
+
+  return {
+    isSpoilerHidden,
   };
 }
