+++ projects/client/src/lib/sections/media-actions/mark-as-watched/useMarkAsWatched.ts
@@ -40,11 +40,7 @@ export function useMarkAsWatched(
   const { user } = useUser();
   const { invalidate } = useInvalidator();
 
-  /**
-   *  TODO: implement some sort of in-memory cache for result of action response
-   *  that will allow to show optimistic UI updates while history is being updated
-   */
-  const isWatched = useIsWatched(props);
+  const { isWatched } = useIsWatched(props);
 
   const markAsWatched = async () => {
     const current = await resolve(user);
